Removed redundant code in Player::SendLoot + Silenced a clang warning
Former-commit-id: 929941af6ff4dc85ea309855dd158d42c58f7b79

+++ src/game/Group.cpp
@@ -1694,7 +1694,7 @@ void Roll::targetObjectBuildLink()
 
 void Group::SetDifficulty(Difficulty difficulty)
 {
-    ASSERT(difficulty < MAX_DIFFICULTY);
+    ASSERT(uint32(difficulty) < MAX_DIFFICULTY);
     m_dungeonDifficulty = difficulty;
     if(!isBGGroup()) 
         CharacterDatabase.PExecute("UPDATE groups SET difficulty = %u WHERE leaderGuid ='%u'", m_dungeonDifficulty, GUID_LOPART(m_leaderGuid));
